"10am. I am up. Yesterday's vacation did me a lot of good. I know I said I would do it for an entire week, but I already feel like resuming.
Yes, it will be impossible to work for the same goals as before. Yes, classical logic is my enemy and is worthy of contempt. Yes, the constructive route will have to be traveled - without some posers trying to stab it in the back.

No, the time for that last part is not now. As much animosity I've gathered towards the mainstream lines of thinking, my real enemy is the low information value of ML experiments.

My programming style has been to go for stabs, but I need to change policies and go for explosives from here on out. Fundamentally, anything constructive math will give me, I can also get using inspired applications of randomized testing.

10:05am. So for the first time in 5 years, not only will I change specific goals, but the general direction I am going in.

It is time to get good.

I do not know whether there is cause to make Spiral, but shitty languages prospering do a lot damage long term and so I should do it as a service to both myself and humanity.

Once I get Spiral up to par, I will be able to tie up the low style and get the full power of what is possible for humans in the domain of programming.

Maybe going for Spiral is not the right choice, but going for Python or any other mainstream language is not the right choice either. So I need to do this. Hopefully it will give me some fame points that in the worst case I will be able to cash in somehow. If I am lucky and the time turns out to be on my side I won't need to, but it never hurts to plan for the worst.

10:10am. Right now it feels I can resume work on the parser lightly. I am going to aim for a few hours a day while I am still in my vacation period and then escalate from there.

In fact, putting a few hours a day might be the ideal pace. It got me that national championship win in the 00s, so it is not like it is an outright mistake.

10:10am. Let me do a little now. I want to see what I can do."