"5:45pm. Let me resume. I want to do the last two patterns. I'll see whether I want to do the outermost parser after that.
```
| PatUnion of Pattern * Pattern list
```

Hmmm...

I am considering that instead of having the union pattern be `a b c`...I should instead make it `A b c`. A variable starting with a big letter would be a union constructor, and the same thing being in the pattern would be an union eliminator.

That way I could have `Some of a | None` instead of `Some of a | None of ()`.

It bothers me that this is not the case in the current design.

```
type Op =
    // Type
    | TypeExpr

    // Unsafe casts
    | UnsafeConvert

    // StringOps
    | StringLength
    | StringIndex
    | StringSlice
```

Would it suck if the union types here all had to have `of ()`?

I should do this. This would be the most elegant.

And on the type level I can reserve the big name variables for nominal types.

5:55pm. To get this effect, I think I will have to modify the tokenizer. I do not feel like that right now.

Let me put the union pattern aside for now and I will just deal with the `when`.

6pm. Ah, I had it. I just forgot to apply the `s`.

Ok.

6:15pm. Ok, I did refactor the tokenizer.

```
| RawTVar of string
```

That reminds me, this really should be `VarString`.

```
let var = (small_var <|> big_var) |>> RawTVar
```

The difference between big and small vars is not really pronounced on the type level.

And...since I've finished all the prereqs let me finally do the union pattern.

```
let pat_union pattern = pipe2 big_var (many ((small_var |>> PatVar) <|> rounds pattern)) (fun a b -> PatUnion(a,b))
```

6:20pm. Perfect. Everything is absolutely perfect.

With this, I've got the patterns down.

6:25pm. I've got the flow back. I will just do it slowly and steadily like this and at some point v0.2 will just be here.

I'll call it a day here. It is time for fun."