"10:15am. Really slept long, but this is supposed to be my vacation so whatever.
Let me rant a bit.

A very abridged version of my learning trajectory in 2019 went NNs -> CFR -> [this](https://stackoverflow.com/questions/59263530/how-to-prove-that-the-halving-function-over-positive-rationals-always-has-an-exi/59304233). It is really at that last point that I started to accept some of my doubts about formal theorem proving. Unlike classical logic which is pretty much a direct enemy of reason and understanding, constructive logic might have something to say about actual reality.

Nonetheless, PLL's answer is really quite apart from my simple model of imagining a number, imagining a bigger number and then repeatedly halving it until it is smaller than the first one. And then mentally verifying that it holds for all cases.

I can't shake off the feeling that while PLL's answer would get me closer to a formal proof, it would actually get me further away from understanding.

And this is what I am going to accept.

Formal proofs might be programs, but they are not understanding. The actual, simple, unproven program that I made is actually closer to understanding than anything else.

10:25am. This is why I suffered for so long, and why I am going to banish this weight looming over me.

I made an extreme effort to get good at theorem proving for a very middling return. I really should take that as an indication of **something**.

10:30am. Just what can I realistically expect understanding to be for CFR, and for deep learning? What can I expect to see? If there are proofs about it, would wrapping them in some bigger program really provide the resolution that I seek?

The doubts eat away at me until I can bear it no longer.

2019 really was a tragedy. My belief in math really got shattered.

I watched Wildberger - that bastion of lucidity's videos, and wondered: just how much effort would it take to do the equivalent of what he is doing there for deep learning and for CFR, the things I am actually interested in? He says that pretty much nothing is known about higher dimensional geometry and my heart sinks.

Deep down, I suspect the answer - literally no human can do this. If those much better at math than me cannot do it, then maybe the answer is not necessarily to just wait 10-1000 years until somebody gets it.

10:35am. I am going to pass the keyboard to the me who understood the self improvement loop from the inside.

I am going to start playing God and start treating ML itself as both science and understanding.

In the future, there will be better algorithms, but I've discovered many principles for how to build robust models which I've yet to really apply as I've considered them to be hacks.

So far I've been treating deep learning as if it were something to be debugged. If I focused harder and dove deeper into it, I would get the necessary understanding to see where the steps forward lie. This methods which was very effective for me for discrete algorithms, utterly falls apart for continuous ones.

10:40am. Right now, there are many worse programmers than me who nevertheless apply the correct principles when it comes to ML while I do not.

I am going to make these last few entries that last word on this. I am going to stop talking and thinking about math.

It is time that I get on with the policy of transcendence.

I did my responsibility towards mathematics. Now it is time I ditch the training weights. I've been tacking them on and on until I could barely move. I had enough of this burden of confusion.

It is time little by little to bring out my full power. I am not going to hold back anymore.

11:05am. The language will be done eventually and my work will slip into more important parts.

Right now, I feel that there is no real need to hurry. It is enough to take some time to heal and imagine how my new principles will work for me. I need to learn to enjoy the future that is coming. I need to indulge a little in it, in order to create the right motivation.

I need to move from being a character defined by my failures, to a character defined by my victories once more.

11:10am. The fundamental reality is that there exists 4-5 layer neural net out there somewhere that could make me millions of dollars in poker or something else.

My expression of understanding from here on out will be to find it. My understanding will be the application of correct principles and the fusion of science and learning into a single program. That is what my new notion of understanding will be.

11:15am. This whole episode really does remind me of the time I was bad at roguelikes. I would create a character, proceed to play it like Diablo and then promptly get killed. Before I became a trader and learned patience, I could never really slip into the right strategic mindset.

I want it. I want to find my power. The expression of understanding is the search itself. ML is not like a regular program. It is a living thing. I do not need to understand search, but do it and interact with it instead. That is what understanding is.

11:20am. This desire of wanting power is the most natural thing in the world. But to achieve good results it is really important to channel it in the right way. Because only if you have the principles aligned with the needs of reality can you create the right flow for both work and learning.

Let me take a break here. It makes no sense to start now.

Instead let me have food, do the chores and then I will do some programming when I feel like it.

That is the way to go for the time being. I have my allowance.

Today, I think I will aim at least to get the outer and the inner patterns out of the way. I've been working on some design bits on the side. This shouldn't be too hard to resolve."