"1:15pm. I am really slacking right now. Let me finish this thing I am reading, do the chores and then I will resume work on the parser.
I am really having fun at the moment. I got up late so I decided to get breakfast immediately today.

1:45pm. Ok, chores. Then I'll get some work in.

2:10pm. It is finally time to start.

Damn, I do not feel like it, but it is just a couple of hours. Nothing I cannot bear.

Now where do I start? I know that I need to do the foralls and the let statements.

2:15pm. What should the syntax be?

`inl f forall a. (x : a) : a = x`

Hmmm, I've been thinking of making the constraints...

`inl f forall a. con a => (x : a) : a = x`

I like this the most of all.

`inl f forall a. (=> con a) (x : a) : a = x`

I've been thinking of this variant for contraints as well as it would be easier to parse, but I do not like it as much.

Also, maybe I should deviate from the F# notation a bit and make it...

`inl f forall a. con a => (x : a) -> a = x`

But this would clash with expressions.

`inl f forall a. con a => (x : a) : a = x`

I like this way the most of all.

But now, there is no way around it anymore. If I want to go down this route, I am going to have to do backtracking parsing.

2:30pm. I am thinking right now.

2:35pm. Ok, I am going to leave constraints for when it is time to do typechecking. That can wait till then.

Ah, yes. I've also completely forgotten to deal with `:`. I am going to have to deal with that as well.

2:50pm. Hmmm...

I need to think about this a bit more. Let me take a short break here."