"10:05am. I went to bed pretty late, but I slept well. Well, it is time to slack. I do not really feel like starting now. It is my vacation and everything, I'll put in the effort later. I've found the English rip of Infinite Dendogram vol 2 and I want to read it. LHscans picked it up as well, so maybe we'll see the other volumes translated in a timely manner as well.
The threads on /a/ about it are always a joy to lurk in.

I saw a while ago that Himenospia will be ending soon. That is really lame. There was a lot of foreshadowing, but that means nothing now - no world war, no shadow war, no human extinction, no nothing. Just the MC licking the floor of her jail cell. Vol 1 of Himenospia was beyond amazing, but now I expect nothing of it.

10:10am. *ahem* Much like I do not expect anything of the field of ML right now.

In fact, that has always been the case. I've always been a true believer in its long term potential, but in the short term I've opted to cast many nets as an expression of my skepticism. I've channeled my skepticism into learning and trying to reason through the algorithms I've looked into. That desire is completely broken now.

Now, I will channel my skepticism through the use of evolutionary methods.

Much like I've dropped type theory in favor of mastering the low style, I will also drop higher order optimizers like KFAC in favor of various tricks. I will use my 'reversibility' ideas to separate out the RL from the SL aspects and that will make KFAC irrelevant. I won't be able to use it smoothly with reversibility anyway and it is just going to be dead weight there.

Also it is incompatible with some of newer tricks that I want to use.

This kind of tends to be the usual outcome with fancy things. You bet on Zap and then realize it does nothing for CFR. You bet on KFAC, and then discover the principles of reversibility. You aim for understanding and then far too quickly reach your limits, realizing that the truest expression of understanding are the programs themselves.

10:20am. I am not going to fail again. This time I will play God without irony.

Let me stop here. Rather than ranting, let me vacation. I'll take care of the let statements and whatever else is needed when it is time to start."