"10:25am. I got up 20m ago, and now it is time to chill. I really am enjoying my allowance to slack. Wasting time is truly the ultimate luxury.
1:15pm. https://mangadex.cc/chapter/689289/1

Let me stop here with Hinamatsuri. Let me just go at it for a bit more and then I will start.

1:40pm. Ready Perfectly. Let me finally start.

First with regards to Spiral I've come to a design decision. Right now there is too much confusion about how to handle default values and such. And I am wondering how am I going to port code between the Easy and the Real segments. For example, in the real part `a + b` would be `((+) a b)` while in the easy segment it would be...```((+) `t a b)``` where that t is an implicit type parameter that needs filling.

There is no way that those `+`es can be the same function.

I've deferred thinking about this up to now, but now it is time to come to a conclusion.

And the conclusion is that the two segments necessarily need different scopes, much like the term and the value levels.

1:50pm. I've been imagining the the real segment should just be able to call the easy segment just like that, but I need to rethink that. The default values are just the first crack that I won't be able to smooth out until I plan this out."