"9:50am. I am up and it is Monday. Tog, Dendro (I am really into it), breakfast, chores, then comes programming.
...Oh, it seems a Satanophany raw came out two days ago.

12:25pm. Finally done with vol 2. The novels were each over 300 pages long, I've been reading them like a madman.

Right now let me take a break from that, I really had enough for the time being and need some downtime. It was pretty good while it lasted, but I want to focus on programming for a bit.

12:30pm. Now before I begin, I still have chores to do, and I want to chill a bit before doing them. Let me get to it.

1:20pm. It is time to start this thing.

1:25pm. Now that I've had my fill of entertainment, and am done with the chores, I finally feel like doing some work.

1:30pm. Now the next thing on the list to do is to take care of error printing.

But I before I start that, there is something that bothers me.

```
    | RawTForall of (string * RawTypeTypeExpr) [] * RawTypeExpr
    | RawTInl of (string * RawTypeTypeExpr) [] * RawTypeExpr
```

The type level functions. There is a certain inconsistency in my thinking.

```
nominal real Tensor dim ty = tensor_init dim ty
```

Remember those nominal types whose body is a call to the real segment function?

1:35pm. Last night (like past 12am) it occurred to me that I do not have a smooth way of taking care of this.

I wanted to simplify things for myself by doing explicit substitution (as opposed to using environments) for type level functions, but this would wreck that. Also on the face of it, while `nominal real Tensor dim ty = tensor_init dim ty` seems super straightforward, I find myself confused as to what `Tensor dim ty` is supposed to be.

It is not a type level function. It is not a forall either. So what should it be then?

Back while I was thinking this through, it was super obvious to me once I had the idea what the effect of this should be. Partially evaluate `tensor_init dim ty` and strip its type. Then use that for the body of the nominal type.

1:50pm. But now that I've looked at it again and am thinking what that really entails, I am finding myself unprepared for having to do value level computation at the type level.

2:40pm. Yeah, this is difficult. I am going to have to add the equivalent of `RawType` to the type expressions. And type level functions will be just like regular functions - I am in fact going to have to give them environments and wait till they are fully applied.

```
    | RawTForall of (string * RawTypeTypeExpr) [] * RawTypeExpr
    | RawTInl of (string * RawTypeTypeExpr) [] * RawTypeExpr
```

I am going to get rid of arrays here. I'll split them up into multiple arguments.

Now, it is great that I've decided all this, but I need to think about it some more. Unification will really be a hell to deal with won't it?"