"3:55pm. I am still thinking about it. I've come to a bunch of design decisions.
```
and RawTypeExpr =
    | RawTVar of VarString
    | RawTPair of RawTypeExpr * RawTypeExpr
    | RawTFun of RawTypeExpr * RawTypeExpr
    | RawTConstraint of RawTypeExpr * RawTypeExpr
    | RawTDepConstraint of RawTypeExpr * RawTypeExpr
    | RawTRecord of Map<string,RawTypeExpr>
    | RawTKeyword of string * RawTypeExpr []
    | RawTApply of RawTypeExpr * RawTypeExpr
    | RawTForall of (string * RawTypeTypeExpr) * RawTypeExpr
    | RawTInl of (string * RawTypeTypeExpr) * RawTypeExpr
    | RawTUnit
    | RawTPos of Pos<RawTypeExpr>
```

4:30pm. ...I am still thinking about it.

I am starting to change my mind about having higher ranked types in the partial evaluator. I am going to need them.

There is something very important that I have just realized now.

```
typecase x with
| a * b -> ...
```

What I hadn't realized is that when `a` and `b` are not bound, the above is in fact equivalent to...

```
typecase x with
| forall a b. a * b -> ...
```

I can have the prepass convert `a * b` to `forall a b. a * b`, but I won't be able to do anything in the peval stage unless I can evaluate such an expression.

```
and RawTypeExpr =
    | RawTVar of VarString
    | RawTPair of RawTypeExpr * RawTypeExpr
    | RawTFun of RawTypeExpr * RawTypeExpr
    | RawTConstraint of RawTypeExpr * RawTypeExpr
    | RawTDepConstraint of RawTypeExpr * RawTypeExpr
    | RawTRecord of Map<string,RawTypeExpr>
    | RawTKeyword of string * RawTypeExpr []
    | RawTApply of RawTypeExpr * RawTypeExpr
    | RawTForall of (string * RawTypeTypeExpr) * RawTypeExpr
    | RawTInl of (string * RawTypeTypeExpr) * RawTypeExpr
    | RawTUnit
    | RawTPos of Pos<RawTypeExpr>
```

Let me paste this here once more. Simply put, it is not realistic for the partial evaluator to not be able to handle higher ranked types. Just how am I going to evaluate `RawTypeExpr` then?

5:25pm. I swear I haven't been slacking today. Though I've barely written a line of code, I've actually generated a lot of flow just sitting here and thinking.

7:05pm. Let me finally call it a day. I am still thinking about it.

I really am quite confused.

No, I can't allow type variables in the partial evaluator after all.

Right now I am thinking in splitting the type system into internal and external types. This would allow me to do unification in typecase and pattern matching.

I really do need to think about it more.

```
inl (a : q),(b : e * r),(c : t) = ...
```

I am really confused as to how binding for type variables should work.

If I allow value level computations in type signatures right there I lose the ability to do unification. And I would very much like to get rid of the type get pattern.

7:25pm. I am still thinking about it.

...I think I am starting to converge on a solution. Internal types for annotations and join point returns and external types for typecase and pattern matching.

The main difference between the two is that internal types will be the true type - meaning I will have them execute the program for real nominals and such, while the external types will just be a description. I do not need to run any value level computations in them and one big difference is that I will allow type variables in them.

I changed my mind on the foralls, but I am definitely going to need type vars.

7:30pm. Hmmm...if I am going to have external types in pattern matching - what about boxing?

7:40pm. Ah, for fuck's sake.

Everything is broken. Shit.

Well, it was fine yesterday. The reason it is broken now is because I decided all of a sudden that I want to have a smooth way of binding the type of an expression.

But regardless, I am going to need external types just for typecase. I won't be able to do unification otherwise.

And for the typechecking phase as well.

7:45pm. Let me just stop here. I need to turn off my brain for a while.

7:50pm. Seriously."