"9:30am. I am up. Yesterday after I stopped, the whole pattern matching issue had me so disturbed that I could not really enjoy myself so I went to bed (on time) instead.
Really, it is amazing that I wasn't even aware of the external/internal split.

9:45am. Ah, whatever. I think I am getting a grasp on this now.

TODO: `inline_` the union pattern.
TODO: Add boxing annotation to `inl`.
TODO: Extend syntax in module open to support renaming.
TODO: Remove the forall restriction in the parser.

Here are my todos from yesterday. I'll get on them later. Right now I feel like continuing my vacationing indefinitely.

I'll up my efforts once I finish the first half of the language. And if not, I am going to up it when I am done with it. Once I catch up where I was in 2018 and start breaking new ground, that will motivate me to really give it my all.

It is the final stretch here till the Singularity.

Spiral v0.2 is going to be the most I will ever contribute to the global pool as a human so I intend to do a good job. Beyond that, any contribution worth making would involve even more ground up work that I do not have the time or power to do.

I did not think it would be like this. I did not think I would work on a language for so long, or that in the end I would end up betting it on all on evolutionary algorithms as science. My powers as human are just so limited. I do not want to be a monkey, but going even a single level beyond just takes such vast effort that the level beyond that is completely out of my reach.

It would all be so easy for a superhuman, but I myself have to constantly face my own limitations.

Nevertheless, this will work. The story is am trying to spin makes sense based on the circumstance. I gave it my all and as a result, I can definitely see more than I did in 2015. The last 1.5 years did not fundamentally break my plan. I just need to refine it a little.

I will do Spiral v0.2 and hopefully some better hardware will come out in this decade that will actually allow me to do large scale RL. Better algorithms will be found eventually as well.

For the sake of power, my life and the future, I will do this. I swear it on my desire.

I will elevate evolution to a science."