"This was a good rant, but now let me slack.
https://vdare.com/articles/eugenic-or-dysgenic-would-world-war-iii-raise-or-lower-iqs

War being eugenic and dysgenic is a question I've been wondering about for a while. Generally, I would assume it would be eugenic if the weak get genocided, but genocide is not in fashion these days. The US might have laid waste to the middle east, but even so it only really tanked their technological base rather than nuked the population. For the eugenic effects to manifest, the fight needs to be to the atrocious, bitter end. It won't happen.

That story which refuses to play out in the age cameras has been pushed to the individual actors.

10:10am. 'Survival of the fittest is survival of the smartest in primitive societies.' Meh. It is true, but this was fairly obvious to me already. I still keep up with these kinds of articles, but rightwing media is sorrow itself.

They talk about intelligence, but if they were really serious they would bet it all on AI research rather than engage in the general kind of larping that they do.

Even 'rationalists' (which aren't right wing in general) succumbed to the sweet lure of alarmism. Why do real work when you can jack morality itself to do your bidding? This is what 'smart' people invest in these days.

Imagine a world when 'waking up' meant 'working on AI'. Where technological progress was backed by adherents following the axiom of power. Where intelligence is not a matter of words, but of battle. A world where the glass of intelligence is not just slowly sipped, but quaffed all the way without hesitation.

In that world, I do not think I could win as there would be smarter, more capable versions of myself everywhere, but I would not have this pervasive sense of loneliness that I do now.

Looking at the world, it is more of the case the it is drawn into the maw of Singularity, rather than actively trying to move towards it. The right is sorrow, the left is nothing. Just where is one to find hope, other than one's own effort? The world is asleep and even those working on AI are not yet awake.

The work on AI is not really work in the truest sense of the word, until those working on it can think of it as alive. It is easy to fiddle with evolution like they are doing now. It will get harder once their reality starts getting invaded and progress necessitates becoming a part of that process. At that point the people will start dropping out of the race and the excesses will get drained both from ML and science itself.

12:25pm. Done with breakfast. Let me play for a bit more, do the chores, and then I will resume.

1:50pm. Done with chores. It is time that I get serious.

Yesterday I stumbled a bit, and today rather than moving on, I am going to need to go back a bit and do some cleaning up. My flow got disrupted, and I need to go from the top once more to see if everything is accounted for.

1:55pm. I won't do anything big today, let just go through the TODOs.

TODO: `inline_` the union pattern.
TODO: Add boxing annotation to `inl`.
TODO: Extend syntax in module open to support renaming.
TODO: Remove the forall restriction in the parser.

I put in the forall restriction in the parser directly, but now that I have constraints and dependent constraints, and who knows what else in the future, it would better to leave dealing with these kinds of error in partial evaluator.

2:05pm. Let me commit this and I will get to work. Right now I am getting distracted by stray thoughts."