"9:40am. I am up. What a night, the less I speak of it the better. Thankfully I will forget it soon.
It is time to chill.

No matter how good Dendro is it really is just fiction. For fiction to take the next evolutionary step, eventually the authors will have to ask the pertinent question of how to take the power from the worlds inside to the outer ones.

The Dendro author is really smart for doing the setup that he did - the MC can essentially respawn and has game-like superpowers, so the main tension is who gets his way and the tians are the ones who have their lives at stake. To me this makes the story more credible. It is not like in regular shonen the main character can just die - they are always protected by plot armor. With this setup there is no (as much) need for that so the story has more integrity as result.

9:50am. But eventually the understanding that the game is just a parasite on the players will dawn.

Overcoming this part is as what I see as my own responsibility.

9:55am. Well, let me waste time as usual, and then I will get my start. Today I will aim to get the first part of the prepass done - the inlining of globals.

There is just so much work to be done, that I won't set any deadlines lest I burn out. The Inspired would be able to do this 24/7 until they are done, but I am going to play this game as what I am."