"2:55pm. Let me resume. Where was I?
```
    | RawGlobalInl of int // Has no free vars. Uses global dictionaries during peval for inlining of recursive occurrences.
    | RawGlobalInlInlined of VarString * RawExpr // Has no free vars.
```

I am not going to need this. I've been debating in myself whether to replace `| RawGlobalInl of int` with `| RawGlobalInl of string`, but the right answer is to just take it out completely.

```
    | RawModule of Map<string,RawExpr>
    | RawModuleOpenInlined of Map<string,RawExpr> * on_succ: RawExpr
```

This is as well.

```
    | RawTInl of (string * RawTypeTypeExpr) * RawTypeExpr
    | RawExpr of RawExpr
```

Among the raw types, let me take out these two.

I can't deal with it right now.

3pm. Ok, focus me. I can chill later.

First of all, how about I bring the `Expr`s back?

3:10pm.

```
// Globals
let mutable tag_prepass = 0
let private tag () = tag_prepass <- tag_prepass + 1; tag_prepass
```

I admit, I completely forgot that Exprs had these tags for easier comparison.

And until I reminded myself just now, it slipped my mind that I was doing the free vars and stack size calculations interleaved with renaming.

```
let prepass x =
    let free_vars_and_stack_size_dict = Dictionary(HashIdentity.Reference)
    let pattern_dict = Dictionary(HashIdentity.Reference)
    let renaming_dict = Dictionary(HashIdentity.Reference)
```

I really, really am getting massive mileage from understanding reference equality.

V0.2 is going to be the version of Spiral where I hammer all of those lessons home.

Now...

3:30pm. I am thinking about it. I am in the flow, but I do not feel like coding just yet.

I was thinking about the inlining, but it had not occurred to me that I would need to merge it with the prepass, so it took me a little off guard.

3:55pm. I am still thinking about it.

I realized just now, arrays are going to have to be nominal types.

Right now, what I am thinking about are how the prepass should be done for the raw typed expressions.

Things like higher order type functions, arrays and primitives.

How do I do inlining of them.

I had not thought about it yet, but now is the time to decide how to do type inbuilts. Should types have ops?

4:30pm. Ok, it turns out there is a lot to think about even before I put in a single line down.

One thing I am also thinking about is getting rid of the builtins.

Instead I could put them directly into the map myself.

One thing I would also want to do is do some inlining in the prepass directly.

For example `a + b` to `!Add(a,b)` where applicable. The way this would work is that I'd have...

`(inl a b -> !Add(a,b)) a b`. This I could reduce in many cases during the prepass into something like `let a = a in let b = b in !Add(a,b)`. And this would speed up partial evaluation by a decent amount.

But there is no shame in leaving that for later.

I can always plug that in when everything else is set.

I should not make it a priority while I am still trying to get the language off the ground.

Later when I have the language working and the test suite operational, then I can do this.

4:40pm. Since it does not seem like I will be writing any code today, let me do a little thing before I leave. Rather than just sit here in this chair like I have for the past 3 hours, it would be more comfortable to think while I am taking a bath.

```
    | RawTInl of (string * RawTypeTypeExpr) * RawTypeExpr
    | RawTPrim of PrimitiveType
    | RawTArray of RawTypeExpr
```

I put the type function back and added these two babies. I am really not going to get far without at least the primitive type and arrays.

For other kinds of types, I can consider that question later.

Once I do the fundamentals they won't really be a problem.

4:50pm. Now let me commit here. I probably just call it a day when I get back so most likely this will be my last entry today.

Without a doubt, today did warm me up towards the prepass quite a bit. I'll start hammering it out tomorrow. Unlike the parser, the prepass itself will be relatively straightforward to deal with. It should not take me longer than a few days.

And when I am done with it comes partial evaluation. That is the home stretch.

The code generator itself won't have much in the changes to speak off unlike the other phases.

Most likely, by the time this month is over I will have already started testing.

And on February that will be when I will add the rest of the features and start work on typechecking itself.

That part will take me a few months, but once I've mastered it, the language will be complete.

What I will be doing at that point is two things: moving on to putting in editor support by writing a VS Code plugin, work on making the error handling and reporting graceful, and actually writing the programs in Spiral itself. Plus some bells and whistles like catching of intermediates.

By the end of the year I should be making good use of Spiral, and by the end of next year not only will I have caught up to par with where I was in 2018, but I will have implemented all the ideas that I have currently. That includes those RL agents.

5pm. 2022 might not mark the start of the Singularity, but it will mark the start of myself as a fully fledged programmer capable of using ML to good effect.

In 2018 and 2019 I pretty much beat myself. I will not let myself be clowned again like that time.

Classical logic might be the enemy of understanding and reason, but attacking it won't be constructive. Instead what I will do is attack the very conception of understanding itself. That should be much more profitable. I will bring out the dark arts that gave me the original inspiration to start this quest.

In 2019 I've made absolutely sure - it is not that my math skill being too low is the main problem. I can get math proofs when they are computational in nature and the subject matter is tractable. The issue is that there are very simple algorithms which are essentially intractable and the ML community is spewing bullshit everywhere.

Previously, I regretfully skipped the proofs as I read papers, but this time I will take a piss on them as I move past them.

Ultimately, as a general method for building understanding, writing programs about programs is highly suspicious at best, and at worst fraud (when done axiomatically.)

Randomized testing is the guiding light. If randomized testing is right then ML itself is the main principle of science.

5:10pm. In order to make progress, I need to get rid of the monkey mentality of looking for the right algorithm to plug in and start thinking of my programming as an expression of understanding itself.

Whether it be DL or CFR, I won't be able to reduce them to something more primitive. They themselves are primitives that I need to use to create a language. Not a programming one, but my own personal, internal one.

That is what it means to make the machines learn.

I will end this."