"9:30am. I am up.
10:50am. Done with vol 7. Let me have that breakfast.

12:40pm. I am nearly a third of the way into vol 8, but let me stop here. It is time for me to mentally start preparing for the programming day. It is also time I deal with the chores.

2:10pm. Today's chores took a bit more than I thought they would. Let me catch my breath for 10m and then I will start.

2:20pm. Let me start.

What I need to do now is really start the prepass.

I just need to do this thing. I feel lethargic just thinking about it.

It is not that any of the pieces are hard, but as is usual at times like these the main problem is gathering momentum.

Still if there is something that I've learned during my long years of programming that would be - if the pieces are there, then it is possible to complete the entire thing.

2:25pm. So focus me, focus. The pieces are all there.

Right now when I think of ML, the pathways that I use to create Spiral are the ones that respond.

This is not a ability that the average guy off the cubicle has. The ML guys certainly do not have it. So I have my own edge. I just need to draw it out.

Just one more time.

I've been saying this like a broken record for the past 5 years, but it is not like 5 years is that long of a time.

In order to win a game, you just need to start making the right moves.

Now that damn prepass.

Let me just focus on annotating the all the functions with their free variables and stack sizes. This is something that I've done before - this time what I need to do is also take note of the type variables.

This kind of simple task should serve me well for what will come after that.

2:30pm. Still...damn.

Yeah, free vars and stack sizes are not difficult by themselves. Neither is the prepass.

But what do I do about inlining?

Inlining is supposed to be done during the prepass (so I do not have to keep around the extra info during caching)

2:40pm. I have no idea.

Hrrrmmmm...Just like yesterday I am going to have to dedicate some time to thinking about this.

There are 3 pieces to the prepass:

1) The free vars and stack size calculations.
2) The renaming and conversion from raw expressions to the array friendly ones that the peval needs.
3) And now inlining of globals.

3:10pm. I am still thinking about it.

3:50pm. I am still thinking about it.

The whole inlining thing being in the last place in the order really throws a wrench in my plans.

4:45pm. I've actually been thinking about it all this time. I nearly have it.

I am going to switch to an alternative naming scheme that will make it unecessary to do things in two passes. This will simplify things.

But the typecase bindings and global inlining makes things complicated and I haven't been able to figure out how to push that through completely yet.

Of course I know ways of doing it, but I haven't settled on the right one exactly.

4:50pm. I just need to go a little bit further in my reasoning - just how do I pass globals around during the prepass?

5:25pm. Ok, I think I've pretty much 95% settled on the new architecture for the prepass.

In fact, I had this idea last time, but decided to go with the more complicated scheme that v0.1 has for some reason.

Now that it has come to this, I need to go all out.

5:30pm. Ok, I figured out one more detail. That is 98% where I am at.

The last 2% is just me wrapping up essentially.

5:35pm. Let me wax the idea for a while longer and then I am going to call it a day.

5:40pm. Ok, I am done.

Here is what I am going to do the basic idea is simple...

```
inl x -> let y = 2 in b + c + x + y
```

Given a function like this, I am going to tag free variables in the negative direction `b -> -1`, `c -> -2` while the bound variables will go in the positive direction `x -> 0`, `y -> 1`. This will make a second of renaming unecessary.

As for the general architecture....

```
let prepass (global_type : Map<string, ExprT>) (global_value : Map<string, Expr>) =
    let rec loop (free_vars_type : Dictionary<string, ExprT>) (free_vars_value : Dictionary<string, Expr>) (vars_type : Set<string>) (vars_value : Set<string>) (local_type : Map<string, ExprT>) (local_value : Map<string, Expr>) = ...
```

5:55pm. Writing this out like this is making me realize that it made more sense in my head.

But...an idea is occurring to me.

Maybe it would not be bad to get rid of maps completely.

Remember that one test in v0.09 which gave me compilation issues I could not get rid of?

```
//let speed3 = // Minus the startup, this takes 0.05s to peval versus 1.5s for the previous version of the compiler.
//    let code =
//        let var i = sprintf "var_%i" i
//        let bnd (a, b) = sprintf "inl %s = %s" a b
//        let vars = [|0..900|] |> Array.map var // Any more than this and it will stack overflow.
//        let bnds =
//            vars |> Array.pairwise |> Array.map (fun (a,b) -> b,a)
//            |> Array.map bnd |> String.concat "\n"
//        let adds = String.concat " + " vars
//        String.concat "\n" [|bnd (var 0, "dyn 0");bnds;adds|]

//    "speed3",([] : Module list),"Does the linear sequence of bindings get compiled in linear time?",code
```

It is difficult to understand what it does just by looking at it now, but as far as I remember, this is a simple program with like 900 different let statements in a row.

Propagating all the free variables takes a crazy amount of time and the main problem is that the immutable data structures are just slow to deal with this. I think the union operation was the main culprit.

```
let prepass (global_type : Map<string, ExprT>) (global_value : Map<string, Expr>) =
    let rec loop (free_vars_type : Dictionary<string, ExprT>) (free_vars_value : Dictionary<string, Expr>) (vars_type : Set<string>) (vars_value : Set<string>) (local_type : Map<string, ExprT>) (local_value : Map<string, Expr>) = ...
```

It seems complicated at first glance, but what if I used a dictionary and cps'd the program so that the adds get removed as soon as the let statements go out of scope.

That would resolve the issues with if statements.

```
(vars_type : Set<string>) (vars_value : Set<string>)
```

As for these two, well, I really need `vars_type` in order to insert the bindings into typecase.

Very well, I guess I'll just keep `vars_type` around for that then.

The missing variables error I can just calculate from the free variables in the global statements.

`(global_type : Map<string, ExprT>) (global_value : Map<string, Expr>)`

These two I will turn into dictionaries as well. No need for immutable maps here.

6:10pm. Let me stop here.

I think the dual naming scheme and the CPS'd dictionaries idea are what I was really looking for. I am going to absolutely crush the prepass this time and forge it to perfection. I admit, before now it never even occurred to me to consider removing things from a dictionary, but it is a great idea.

6:20pm. Oh, yes. I am going to need `vars_value` if nothing else, but just to avoid inlining the globals incorrectly.

Well, I'll figure it out tomorrow as I go along."